Modified files for CAMB and Liklihood
The power_tilt file is in the CAMB folder
The camb, model, initial power are all files in the pycamb -> camb folder 
The Likelihood is the file to do the fisher forecast. It needs to be in the pycamb. 
The rough file contains the notes for the LIklihood file - in particular what we are using for the fisher forecast